+++ flink-runtime/src/test/java/org/apache/flink/runtime/state/MemoryStateBackendTest.java
@@ -91,44 +91,6 @@ public class MemoryStateBackendTest extends StateBackendTestBase<MemoryStateBack
 	public void testMapStateRestoreWithWrongSerializers() {}
 
 	@Test
-	@SuppressWarnings("unchecked")
-	public void testNumStateEntries() throws Exception {
-		KeyedStateBackend<Integer> backend = createKeyedBackend(IntSerializer.INSTANCE);
-
-		ValueStateDescriptor<String> kvId = new ValueStateDescriptor<>("id", String.class, null);
-		kvId.initializeSerializerUnlessSet(new ExecutionConfig());
-
-		HeapKeyedStateBackend<Integer> heapBackend = (HeapKeyedStateBackend<Integer>) backend;
-
-		assertEquals(0, heapBackend.numStateEntries());
-
-		ValueState<String> state = backend.getPartitionedState(VoidNamespace.INSTANCE, VoidNamespaceSerializer.INSTANCE, kvId);
-
-		backend.setCurrentKey(0);
-		state.update("hello");
-		state.update("ciao");
-
-		assertEquals(1, heapBackend.numStateEntries());
-
-		backend.setCurrentKey(42);
-		state.update("foo");
-
-		assertEquals(2, heapBackend.numStateEntries());
-
-		backend.setCurrentKey(0);
-		state.clear();
-
-		assertEquals(1, heapBackend.numStateEntries());
-
-		backend.setCurrentKey(42);
-		state.clear();
-
-		assertEquals(0, heapBackend.numStateEntries());
-
-		backend.dispose();
-	}
-
-	@Test
 	public void testOversizedState() {
 		try {
 			MemoryStateBackend backend = new MemoryStateBackend(10);
